@@ -6,7 +6,7 @@ import 'dart:typed_data';
 /// Read [stream] into a String.
 ///
 /// Defaults to [utf8] if no [encoding] is given.
-Future<String> readAsString(Stream<List<int>> stream, {Encoding encoding}) {
+Future<String> readAsString(Stream<List<int>> stream, {Encoding? encoding}) {
   encoding ??= utf8;
   return encoding.decodeStream(stream);
 }
@@ -20,7 +20,7 @@ Future<String> readAsString(Stream<List<int>> stream, {Encoding encoding}) {
 /// (e.g. because the underlying list may get modified).
 Future<Uint8List> readAsBytes(
     Stream<List<int>> stream, {
-      int maxLength,
+      int? maxLength,
       bool copy = false,
     }) async {
   final bb = BytesBuffer();
@@ -44,7 +44,7 @@ Future<Uint8List> readAsBytes(
 Stream<Uint8List> sliceStream(
     Stream<List<int>> stream,
     int sliceLength, {
-      int maxLength,
+      int? maxLength,
       bool copy = false,
     }) async* {
   var total = 0;
@@ -62,7 +62,7 @@ Stream<Uint8List> sliceStream(
 
     while (getBL() >= sliceLength) {
       final bufferLength = getBL();
-      Uint8List overflow;
+      Uint8List? overflow;
       if (bufferLength > sliceLength) {
         final last = buffer.removeLast();
         final index = sliceLength - bufferLength + last.length;
@@ -130,7 +130,7 @@ class BytesBuffer {
   ///
   /// Set [copy] to true if [bytes] need to be copied (e.g. the underlying
   /// buffer will be modified.)
-  void add(List<int> bytes, {bool copy}) {
+  void add(List<int> bytes, {bool? copy}) {
     _chunks.add(castBytes(bytes, copy: copy ?? _copy));
     _length += bytes.length;
   }
@@ -141,7 +141,7 @@ class BytesBuffer {
   }
 
   /// Concatenate the byte arrays and return them as a single unit.
-  Uint8List toBytes({bool copy}) {
+  Uint8List toBytes({bool? copy}) {
     if (_chunks.length == 1 && !(copy ?? _copy)) {
       return _chunks.single;
     }
@@ -169,7 +169,7 @@ class ByteDataWriter {
   bool _dataEmpty = true;
   ByteData __data = ByteData(0);
   ByteData get _data => __data;
-  set _data(ByteData data) {
+  set _data(ByteData? data) {
     if (data == null) {
       _dataEmpty = true;
       __data = ByteData(0);
@@ -206,13 +206,13 @@ class ByteDataWriter {
     _bb.add(bytes, copy: copy);
   }
 
-  void writeFloat32(double value, [Endian endian]) {
+  void writeFloat32(double value, [Endian? endian]) {
     _init(4);
     _data.setFloat32(_offset, value, endian ?? this.endian);
     _offset += 4;
   }
 
-  void writeFloat64(double value, [Endian endian]) {
+  void writeFloat64(double value, [Endian? endian]) {
     _init(8);
     _data.setFloat64(_offset, value, endian ?? this.endian);
     _offset += 8;
@@ -224,25 +224,25 @@ class ByteDataWriter {
     _offset++;
   }
 
-  void writeInt16(int value, [Endian endian]) {
+  void writeInt16(int value, [Endian? endian]) {
     _init(2);
     _data.setInt16(_offset, value, endian ?? this.endian);
     _offset += 2;
   }
 
-  void writeInt32(int value, [Endian endian]) {
+  void writeInt32(int value, [Endian? endian]) {
     _init(4);
     _data.setInt32(_offset, value, endian ?? this.endian);
     _offset += 4;
   }
 
-  void writeInt64(int value, [Endian endian]) {
+  void writeInt64(int value, [Endian? endian]) {
     _init(8);
     _data.setInt64(_offset, value, endian ?? this.endian);
     _offset += 8;
   }
 
-  void writeInt(int byteLength, int value, [Endian endian]) {
+  void writeInt(int byteLength, int value, [Endian? endian]) {
     switch (byteLength) {
       case 1:
         writeInt8(value);
@@ -268,25 +268,25 @@ class ByteDataWriter {
     _offset++;
   }
 
-  void writeUint16(int value, [Endian endian]) {
+  void writeUint16(int value, [Endian? endian]) {
     _init(2);
     _data.setUint16(_offset, value, endian ?? this.endian);
     _offset += 2;
   }
 
-  void writeUint32(int value, [Endian endian]) {
+  void writeUint32(int value, [Endian? endian]) {
     _init(4);
     _data.setUint32(_offset, value, endian ?? this.endian);
     _offset += 4;
   }
 
-  void writeUint64(int value, [Endian endian]) {
+  void writeUint64(int value, [Endian? endian]) {
     _init(8);
     _data.setUint64(_offset, value, endian ?? this.endian);
     _offset += 8;
   }
 
-  void writeUint(int byteLength, int value, [Endian endian]) {
+  void writeUint(int byteLength, int value, [Endian? endian]) {
     switch (byteLength) {
       case 1:
         writeUint8(value);
@@ -327,7 +327,7 @@ class ByteDataReader {
   bool _dataEmpty = true;
   ByteData __data = ByteData(0);
   ByteData get _data => __data;
-  set _data(ByteData data) {
+  set _data(ByteData? data) {
     if (data == null) {
       _dataEmpty = true;
       __data = ByteData(0);
@@ -337,7 +337,7 @@ class ByteDataReader {
     _dataEmpty = false;
   }
 
-  Completer _readAheadCompleter;
+  Completer? _readAheadCompleter;
   int _readAheadRequired = 0;
 
   ByteDataReader({this.endian = Endian.big, bool copy = false}) : _copy = copy;
@@ -384,12 +384,12 @@ class ByteDataReader {
     }
   }
 
-  void add(List<int> bytes, {bool copy}) {
+  void add(List<int> bytes, {bool? copy}) {
     _queue.add(castBytes(bytes, copy: copy ?? _copy));
     _queueCurrentLength += bytes.length;
     _queueTotalLength += bytes.length;
     if (_readAheadCompleter != null && remainingLength >= _readAheadRequired) {
-      _readAheadCompleter.complete();
+      _readAheadCompleter!.complete();
       _readAheadCompleter = null;
     }
   }
@@ -400,17 +400,17 @@ class ByteDataReader {
       return Future.value();
     }
     if (_readAheadCompleter != null && _readAheadRequired == length) {
-      return _readAheadCompleter.future;
+      return _readAheadCompleter!.future;
     }
     if (_readAheadCompleter != null && _readAheadRequired != length) {
       throw StateError('A different readAhead is already waiting.');
     }
     _readAheadRequired = length;
     _readAheadCompleter = Completer();
-    return _readAheadCompleter.future;
+    return _readAheadCompleter!.future;
   }
 
-  Uint8List read(int length, {bool copy}) {
+  Uint8List read(int length, {bool? copy}) {
     if (length == 0) {
       return Uint8List(0);
     }
@@ -448,14 +448,14 @@ class ByteDataReader {
     return bb.toBytes();
   }
 
-  double readFloat32([Endian endian]) {
+  double readFloat32([Endian? endian]) {
     _init(4);
     final value = _data.getFloat32(_offset, endian ?? this.endian);
     _offset += 4;
     return value;
   }
 
-  double readFloat64([Endian endian]) {
+  double readFloat64([Endian? endian]) {
     _init(8);
     final value = _data.getFloat64(_offset, endian ?? this.endian);
     _offset += 8;
@@ -469,28 +469,28 @@ class ByteDataReader {
     return value;
   }
 
-  int readInt16([Endian endian]) {
+  int readInt16([Endian? endian]) {
     _init(2);
     final value = _data.getInt16(_offset, endian ?? this.endian);
     _offset += 2;
     return value;
   }
 
-  int readInt32([Endian endian]) {
+  int readInt32([Endian? endian]) {
     _init(4);
     final value = _data.getInt32(_offset, endian ?? this.endian);
     _offset += 4;
     return value;
   }
 
-  int readInt64([Endian endian]) {
+  int readInt64([Endian? endian]) {
     _init(8);
     final value = _data.getInt64(_offset, endian ?? this.endian);
     _offset += 8;
     return value;
   }
 
-  int readInt(int byteLength, [Endian endian]) {
+  int readInt(int byteLength, [Endian? endian]) {
     switch (byteLength) {
       case 1:
         return readInt8();
@@ -513,28 +513,28 @@ class ByteDataReader {
     return value;
   }
 
-  int readUint16([Endian endian]) {
+  int readUint16([Endian? endian]) {
     _init(2);
     final value = _data.getUint16(_offset, endian ?? this.endian);
     _offset += 2;
     return value;
   }
 
-  int readUint32([Endian endian]) {
+  int readUint32([Endian? endian]) {
     _init(4);
     final value = _data.getUint32(_offset, endian ?? this.endian);
     _offset += 4;
     return value;
   }
 
-  int readUint64([Endian endian]) {
+  int readUint64([Endian? endian]) {
     _init(8);
     final value = _data.getUint64(_offset, endian ?? this.endian);
     _offset += 8;
     return value;
   }
 
-  int readUint(int byteLength, [Endian endian]) {
+  int readUint(int byteLength, [Endian? endian]) {
     switch (byteLength) {
       case 1:
         return readUint8();
