@@ -8,44 +8,44 @@ import 'upvalue_holder.dart';
 class LuaStack {
 
   /// virtual stack
-  final List<Object> slots =  List();
+  final List<Object?> slots =  [];
   /// call info
-  LuaStateImpl state;
-  Closure closure;
-  List<Object> varargs;
-  Map<int, UpvalueHolder> openuvs;
+  late LuaStateImpl state;
+  Closure? closure;
+  List<Object?>? varargs;
+  Map<int?, UpvalueHolder?>? openuvs;
   /// Program Counter
   int pc = 0;
   /// linked list
-  LuaStack prev;
+  LuaStack? prev;
 
   int top() {
     return slots.length;
   }
 
-  void push(Object val) {
+  void push(Object? val) {
     if (slots.length > 10000) { // TODO
       throw StackOverflowError();
     }
     slots.add(val);
   }
 
-  Object pop() {
+  Object? pop() {
     return slots.removeAt(slots.length - 1);
   }
 
-  void pushN(List<Object> vals, int n) {
+  void pushN(List<Object?>? vals, int n) {
     int nVals = vals == null ? 0 : vals.length;
     if (n < 0) {
       n = nVals;
     }
     for (int i = 0; i < n; i++) {
-      push(i < nVals ? vals[i] : null);
+      push(i < nVals ? vals![i] : null);
     }
   }
 
-  List<Object> popN(int n) {
-    List<Object> vals = List<Object>();
+  List<Object?> popN(int n) {
+    List<Object?> vals = <Object?>[];
     for (int i = 0; i < n; i++) {
       vals.add(pop());
     }
@@ -60,7 +60,7 @@ class LuaStack {
   bool isValid(int idx) {
     if (idx < lua_registryindex) { /* upvalues */
       int uvIdx = lua_registryindex - idx - 1;
-      return closure != null && uvIdx < closure.upvals.length;
+      return closure != null && uvIdx < closure!.upvals.length;
     }
 
     if (idx == lua_registryindex) {
@@ -70,13 +70,13 @@ class LuaStack {
     return absIdx > 0 && absIdx <= slots.length;
   }
 
-  Object get(int idx) {
+  Object? get(int idx) {
     if (idx < lua_registryindex) { /* upvalues */
       int uvIdx = lua_registryindex - idx - 1;
       if (closure != null
-          && closure.upvals.length > uvIdx
-          && closure.upvals[uvIdx] != null) {
-        return closure.upvals[uvIdx].get();
+          && closure!.upvals.length > uvIdx
+          && closure!.upvals[uvIdx] != null) {
+        return closure!.upvals[uvIdx]!.get();
       } else {
         return null;
       }
@@ -93,19 +93,19 @@ class LuaStack {
     }
   }
 
-  void set(int idx, Object val) {
+  void set(int idx, Object? val) {
     if (idx < lua_registryindex) { /* upvalues */
       int uvIdx = lua_registryindex - idx - 1;
       if (closure != null
-          && closure.upvals.length > uvIdx
-          && closure.upvals[uvIdx] != null) {
-        closure.upvals[uvIdx].set(val);
+          && closure!.upvals.length > uvIdx
+          && closure!.upvals[uvIdx] != null) {
+        closure!.upvals[uvIdx]!.set(val);
       }
       return;
     }
 
     if (idx == lua_registryindex) {
-      state.registry = (val as LuaTable);
+      state.registry = (val as LuaTable?);
       return;
     }
     int absIdx = absIndex(idx);
