@@ -4,48 +4,48 @@ import '../number/lua_number.dart';
 
 class LuaTable {
   /// 元表
-  LuaTable metatable;
-  List<Object> arr;
-  Map<Object, Object> map;
+  LuaTable? metatable;
+  List<Object?>? arr;
+  Map<Object?, Object>? map;
 
   // used by next()
-  Map<Object, Object> keys;
-  Object lastKey;
-  bool changed;
+  Map<Object?, Object?>? keys;
+  Object? lastKey;
+  late bool changed;
 
   LuaTable(int nArr, int nRec) {
     if (nArr > 0) {
       // arr = List<Object>(nArr);
-      arr = List<Object>();
+      arr = <Object?>[];
     }
     if (nRec > 0) {
       // map =  Map<Object, Object>(nRec);
-      map =  HashMap<Object, Object>();
+      map =  HashMap<Object?, Object>();
     }
   }
 
   bool hasMetafield(String fieldName) {
-    return metatable != null && metatable.get(fieldName) != null;
+    return metatable != null && metatable!.get(fieldName) != null;
   }
 
   int length() {
-    return arr == null ? 0 : arr.length;
+    return arr == null ? 0 : arr!.length;
   }
 
-  Object get(Object key) {
+  Object? get(Object? key) {
     key = floatToInteger(key);
 
     if (arr != null && key is int) {
       int idx = key;
-      if (idx >= 1 && idx <= arr.length) {
-        return arr[idx - 1];
+      if (idx >= 1 && idx <= arr!.length) {
+        return arr![idx - 1];
       }
     }
 
-    return map != null ? map[key] : null;
+    return map != null ? map![key] : null;
   }
 
-  void put(Object key, Object val) {
+  void put(Object? key, Object? val) {
     if (key == null) {
       throw Exception("table index is nil!");
     }
@@ -58,12 +58,12 @@ class LuaTable {
       int idx = key;
       if (idx >= 1) {
         if (arr == null) {
-          arr = List<Object>();
+          arr = <Object?>[];
         }
 
-        int arrLen = arr.length;
+        int arrLen = arr!.length;
         if (idx <= arrLen) {
-          arr[idx-1] = val;
+          arr![idx-1] = val;
           if (idx == arrLen && val == null) {
             shrinkArray();
           }
@@ -71,10 +71,10 @@ class LuaTable {
         }
         if (idx == arrLen + 1) {
           if (map != null) {
-            map.remove(key);
+            map!.remove(key);
           }
           if (val != null) {
-            arr.add(val);
+            arr!.add(val);
             expandArray();
           }
           return;
@@ -84,17 +84,17 @@ class LuaTable {
 
     if (val != null) {
       if (map == null) {
-        map = HashMap<Object, Object>();
+        map = HashMap<Object?, Object>();
       }
-      map[key] = val;
+      map![key] = val;
     } else {
       if (map != null) {
-        map.remove(key);
+        map!.remove(key);
       }
     }
   }
 
-  Object floatToInteger(Object key) {
+  Object? floatToInteger(Object? key) {
     if (key is double) {
       double f = key;
       if (LuaNumber.isInteger(f)) {
@@ -105,19 +105,19 @@ class LuaTable {
   }
 
   void shrinkArray() {
-    for (int i = arr.length - 1; i >= 0; i--) {
-      if (arr[i] == null) {
-        arr.removeAt(i);
+    for (int i = arr!.length - 1; i >= 0; i--) {
+      if (arr![i] == null) {
+        arr!.removeAt(i);
       }
     }
   }
 
   void expandArray() {
     if (map != null) {
-      for (int idx = arr.length + 1; ; idx++) {
-        Object val = map.remove(idx);
+      for (int idx = arr!.length + 1; ; idx++) {
+        Object? val = map!.remove(idx);
         if (val != null) {
-          arr.add(val);
+          arr!.add(val);
         } else {
           break;
         }
@@ -125,13 +125,13 @@ class LuaTable {
     }
   }
 
-  Object nextKey(Object key) {
+  Object? nextKey(Object? key) {
     if (keys == null || (key == null && changed)) {
       initKeys();
       changed = false;
     }
 
-    Object nextKey = keys[key];
+    Object? nextKey = keys![key];
     if (nextKey == null && key != null && key != lastKey) {
       throw Exception("invalid key to 'next'");
     }
@@ -141,25 +141,25 @@ class LuaTable {
 
   void initKeys() {
     if (keys == null) {
-      keys = HashMap<Object, Object>();
+      keys = HashMap<Object?, Object?>();
     } else {
-      keys.clear();
+      keys!.clear();
     }
-    Object key = null;
+    Object? key = null;
     if (arr != null) {
-      for (int i = 0; i < arr.length; i++) {
-        if (arr[i] != null) {
+      for (int i = 0; i < arr!.length; i++) {
+        if (arr![i] != null) {
           int nextKey = i + 1;
-          keys[key] = nextKey;
+          keys![key] = nextKey;
           key = nextKey;
         }
       }
     }
     if (map != null) {
-      for (Object k in map.keys) {
-        Object v = map[k];
+      for (Object? k in map!.keys) {
+        Object? v = map![k];
         if (v != null) {
-          keys[key] = k;
+          keys![key] = k;
           key = k;
         }
       }
