@@ -3,21 +3,21 @@ import 'lua_stack.dart';
 
 class UpvalueHolder {
 
-  final int index;
-  LuaStack stack;
-  Object value;
+  final int? index;
+  LuaStack? stack;
+  Object? value;
 
   UpvalueHolder.value(this.value) :this.index = 0;
 
   UpvalueHolder(this.stack, this.index);
 
-  Object get() {
-    return stack != null ? stack.get(index + 1) : value;
+  Object? get() {
+    return stack != null ? stack!.get(index! + 1) : value;
   }
 
-  void set(Object value) {
+  void set(Object? value) {
     if (stack != null) {
-      stack.set(index + 1, value);
+      stack!.set(index! + 1, value);
     } else {
       this.value = value;
     }
@@ -25,7 +25,7 @@ class UpvalueHolder {
 
   void migrate() {
     if (stack != null) {
-      value = stack.get(index + 1);
+      value = stack!.get(index! + 1);
       stack = null;
     }
   }
