@@ -77,6 +77,7 @@ class CharSequence {
   }
 
   static bool isalnum(String c) {
+    if(c.isEmpty) return false;
     var code = c.codeUnitAt(0);
     // '0'~'9' or a~z or A~Z
     return code >= 48 && code <= 57 ||
@@ -85,7 +86,6 @@ class CharSequence {
   }
 
   static int count(String src,String ch){
-    if(src == null) return -1;
     if(src.isEmpty) return 0;
 
     var sum=0;
