@@ -5,7 +5,7 @@ class LuaNumber {
   }
 
   // TODO
-  static int parseInteger(String str) {
+  static int? parseInteger(String str) {
     try {
       return int.parse(str);
     } catch (e) {
@@ -14,7 +14,7 @@ class LuaNumber {
   }
 
   // TODO
-  static double parseFloat(String str) {
+  static double? parseFloat(String str) {
     try {
       return double.parse(str);
     } catch (e) {
