@@ -1,9 +1,9 @@
 
 class Userdata<T>{
 
-  final List<T> _data = List(1);
+  final List<T?> _data = List.filled(1,null);
 
-  get data => _data.first;
+  T? get data => _data.first;
 
-  set data(T data)=> _data.first = data;
+  set data(T? data)=> _data.first = data;
 }
